+++ src/main/java/com/nhnacademy/store99/bookstore/order_book/repository/OrderBookRepository.java
@@ -1,10 +1,10 @@
-package com.nhnacademy.store99.bookstore.order_book.entity.service;
+package com.nhnacademy.store99.bookstore.order_book.repository;
 
-import com.nhnacademy.store99.bookstore.order_book.entity.response.LatestBookResponse;
+import com.nhnacademy.store99.bookstore.order_book.response.LatestBookResponse;
 import java.util.List;
 import java.util.Map;
 
-public interface OrderBookService {
+public interface OrderBookRepository {
 
     // 이 쿼리는 각 도서의 주문갯수를 가져옴. 그리고 가장 많이 팔리는걸 처음으로 가져옴.
     // select ob.book_id, COUNT(ob.book_id)  from order_book ob GROUP BY ob.book_id  Order By COUNT(ob.book_id) DESC;
@@ -14,3 +14,5 @@ public interface OrderBookService {
     // select b.book_id, b.created_at  from books b ORDER BY b.created_at DESC, b.book_id DESC;
     List<LatestBookResponse> latestBooks();
 }
+
+
